refactor: move mocks to separate folder
